+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-02 09:02:50
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-07-02 12:42:38
+// Commit: fix(state): fix memory leak by cancelling state subscription
 
 class stateComponentState {
-  final String commitRef = '2454';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-07-02T09:02:50+05:00');
+  final String commitRef = '2465';
+  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
+  final DateTime timestamp = DateTime.parse('2026-07-02T12:42:38+05:00');
 }